move it agent to msic
