+++ hud/agents/misc/__init__.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from .integration_test_agent import IntegrationTestRunner
 from .response_agent import ResponseAgent
 
-__all__ = ["ResponseAgent"]
+__all__ = ["ResponseAgent", "IntegrationTestRunner"]
